refactor: reorganize folders and refactor code
