+++ keeper/src/trade/withdrawal/handle.rs
@@ -5,29 +5,63 @@ use cainome::{
     rs::abigen,
 };
 use starknet::{
-    accounts::SingleOwnerAccount,
+    accounts::{Account, Call, SingleOwnerAccount},
     core::types::FieldElement,
     providers::jsonrpc::{HttpTransport, JsonRpcClient},
     signers::LocalWallet,
 };
 
-use crate::types::SatoruAction;
+use crate::{trade::utils::get_set_primary_price_call, types::SatoruAction};
 
 abigen!(
     WithdrawalHandler,
-    "./resources/satoru_WithdrawalHandler.contract_class.json"
+    "./resources/satoru_WithdrawalHandler.contract_class.json",
+);
+
+abigen!(
+    DataStore,
+    "./resources/satoru_DataStore.contract_class.json",
+    type_aliases {
+        satoru::data::data_store::DataStore::Event as Event_;
+        satoru::utils::span32::Span32 as Span32_;
+    }
+);
+
+abigen!(
+    Oracle,
+    "./resources/satoru_Oracle.contract_class.json",
+    type_aliases {
+        satoru::price::price::Price as Price_;
+        satoru::oracle::oracle::Oracle::Event as Event__;
+    }
 );
 
 pub async fn handle_withdrawal(
     account: Arc<SingleOwnerAccount<JsonRpcClient<HttpTransport>, LocalWallet>>,
     withdrawal: SatoruAction,
 ) {
+    let set_price_call = get_set_primary_price_call(withdrawal.clone(), account.clone()).await;
+
+    let execute_withdrawal_call = get_execute_withdrawal_call(withdrawal, account.clone());
+
+    let _withdrawal_execution_multicall = account
+        .execute(vec![set_price_call, execute_withdrawal_call])
+        .send()
+        .await
+        .expect("Withdrawal execution multicall failed");
+    // TODO: poll transaction status
+}
+
+fn get_execute_withdrawal_call(
+    withdrawal: SatoruAction,
+    account: Arc<SingleOwnerAccount<JsonRpcClient<HttpTransport>, LocalWallet>>,
+) -> Call {
     let withdrawal_handler_address =
         env::var("WITHDRAWAL_HANDLER").expect("WITHDRAWAL_HANDLER env variable not set");
     let withdrawal_handler = WithdrawalHandler::new(
         FieldElement::from_hex_be(&withdrawal_handler_address)
             .expect("Conversion error: withdrawal_handler_address"),
-        account,
+        account.clone(),
     );
 
     let set_prices_params: SetPricesParams = SetPricesParams {
@@ -67,14 +101,8 @@ pub async fn handle_withdrawal(
         price_feed_tokens: vec![],
     };
 
-    let tx = withdrawal_handler
-        .execute_withdrawal(
-            &FieldElement::from_hex_be(&withdrawal.key).expect("Cannot convert string to felt"),
-            &set_prices_params,
-        )
-        .send()
-        .await
-        .expect("Withdrawal Execution Failed");
-
-    // TODO: poll transaction status
+    withdrawal_handler.execute_withdrawal_getcall(
+        &FieldElement::from_hex_be(&withdrawal.key).expect("Cannot convert string to felt"),
+        &set_prices_params,
+    )
 }
